+++ src/main/java/id/co/microservice/currency/currency_service/service/strategy/SupportedCurrenciesStrategy.java
@@ -1,4 +1,4 @@
-package id.co.microservice.currency.currency_service.service.impl;
+package id.co.microservice.currency.currency_service.service.strategy;
 
 import id.co.microservice.currency.currency_service.config.ExternalApiConfig;
 import id.co.microservice.currency.currency_service.constant.CurrencyConstant;
